fix: Ensure correct permissions are set for all files.
File permissions weren't being set after unzipping, which resulted in whatever permissions git gives them - which doesn't always allow write permissions for all users. If the script is run by one user and then by another, the second user may not have write permissions and so cannot unzip the package.

Also, it is good practice to not arbitrarily make files executable. Directories must be (permission to open a directory is dictated by the execution permission) but allowing execution of files we don't fully control is not a good idea.